PCAAXS1: Enable/Disable LVDS transceiver
The LVDS transceiver outputs a pclk even when no data/clock
is generated by the AM335x. This pclk sets the LCD-018 ETM0700G0DH6
into a testmode. As we can not enable and disable the
display over disp_en signal in software we use the power_dwn
of the LVDS transceiver instead.

Also prevent reset of the GPIO bank of the power_dwn pin.

Signed-off-by: Teresa Gámez <t.gamez@phytec.de>