PLFA03: Add enable function for ecap2
The late_initcalls are called form all boards which are enabled.
So when booting a phyCORE-AM335x the late_initcall for ecap2 in the
phyFLEX-AM335x board code will be called, too.
This causes warnings while booting.

Enable the registration of the backlight device in the late_initcall only
when booting from phyFLEX-AM335x.

We do need the late_initcall to prevent flickering while booting.

Signed-off-by: Teresa Gámez <t.gamez@phytec.de>